requirement.txt
pymongo
pandas
python-dotenv
google-cloud-bigquery
google-cloud-bigquery-storage
pyarrow
faker